Added publisher form

+++ src/main/java/com/controller/MainController.java
@@ -63,7 +63,7 @@ public class MainController {
 
     @RequestMapping(value = "/deleteBooks.ajax")
     public @ResponseBody
-    String deleteBook(@RequestParam Long bookId, HttpServletRequest request, HttpServletResponse response) throws ParseException {
+    String deleteBook(@RequestParam Integer bookId, HttpServletRequest request, HttpServletResponse response) throws ParseException {
 
         JSONObject jsonObject = bookService.deleteBook(bookId);
         return jsonObject.toString();
@@ -97,7 +97,7 @@ public class MainController {
 
     @RequestMapping(value = "/deletePublisher.ajax")
     public @ResponseBody
-    String deletePublisher(@RequestParam Long publisherId, HttpServletRequest request, HttpServletResponse response) throws ParseException {
+    String deletePublisher(@RequestParam Integer publisherId, HttpServletRequest request, HttpServletResponse response) throws ParseException {
 
         JSONObject jsonObject = publisherService.deletePublisher(publisherId);
         return jsonObject.toString();
